+++ qabelbox/src/main/java/de/qabel/qabelbox/chat/services/QabelFirebaseMessagingService.kt
@@ -4,16 +4,20 @@ import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
 import de.qabel.qabelbox.QabelBoxApplication
 import de.qabel.qabelbox.helper.AccountHelper
+import de.qabel.qabelbox.reporter.CrashReporter
 import org.jetbrains.anko.AnkoLogger
 import org.jetbrains.anko.info
 import org.jetbrains.anko.warn
+import javax.inject.Inject
 
 class QabelFirebaseMessagingService: FirebaseMessagingService(), AnkoLogger {
 
+    @Inject lateinit var crashReporter: CrashReporter
 
     override fun onCreate() {
         super.onCreate()
         QabelBoxApplication.getApplicationComponent(applicationContext).inject(this)
+        crashReporter.installCrashReporter()
         info("QabelFireBaseMessagingService started")
     }