+++ qabelbox/src/main/java/de/qabel/qabelbox/dagger/modules/ApplicationModule.java
@@ -33,11 +33,16 @@ import de.qabel.qabelbox.identity.interactor.MainIdentityUseCase;
 import de.qabel.qabelbox.listeners.ActionIntentSender;
 import de.qabel.qabelbox.listeners.AndroidActionIntentCastSender;
 
+ormers.ChatMessageTransformer;
+
 @Module
 public class ApplicationModule extends ContextModule {
 
+    private QabelBoxApplication application;
+
     public ApplicationModule(QabelBoxApplication application) {
         super(application);
+        this.application = application;
     }
 
     @Provides